pyproject.toml
[project]
name = "coding-proxy"
version = "0.1.0a1"
description = "A High-Availability, Transparent, and Smart Multi-Vendor Proxy for Claude Code. Support Claude Plans, GitHub Copilot, Google Antigravity, ZAI/GLM."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "ThreeFish-AI", email = "threefish.ai@gmail.com" }]
license = "Apache-2.0"
keywords = ["claude-code", "vibe-coding", "llm-agent", "antigravity", "github-copilot", "glm"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "fastapi>=0.135",
    "uvicorn>=0.42",
    "httpx>=0.28",
    "pydantic>=2.12",
    "pyyaml>=6.0",
    "typer>=0.12.5,<0.22.0",
    "aiosqlite>=0.22",
    "rich>=14.3",
]

[project.urls]
"Source Code" = "https://github.com/ThreeFish-AI/coding-proxy"
"User Guide" = "https://github.com/ThreeFish-AI/coding-proxy/blob/master/docs/user-guide.md"
"Bug Tracker" = "https://github.com/ThreeFish-AI/coding-proxy/issues"

[project.scripts]
coding-proxy = "coding.proxy.cli:app"

[build-system]
requires = ["hatchling>=1.29"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coding"]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "pytest-cov>=6.0",
    "ruff>=0.9.0",
]

# ── Ruff: Linter & Formatter（替代 flake8 + isort + Black + pyupgrade 等 6+ 工具） ──
# 文档: https://docs.astral.sh/ruff/
[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "F",   # Pyflakes（未定义变量、导入错误等 — 捕获真实 Bug）
    "E",   # pycodestyle 错误
    "W",   # pycodestyle 警告
    "I",   # isort（导入排序）
    "UP",  # pyupgrade（Python 语法现代化，安全自动修复）
]
ignore = [
    "E501", # 行过长（由 formatter 处理）
    "E402", # 模块级非顶层导入（__init__.py 文档化分组等合理场景）
]

[tool.ruff.lint.isort]
known-first-party = ["coding"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # __init__.py 允许未使用的导入（re-export）

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ── Pytest 配置 ──
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]
